+++ Changelog.md
@@ -33,6 +33,12 @@
 
 ## Ignition Math 2.x
 
+1. Added OrientedBox
+    * [Pull request 146](https://bitbucket.org/ignitionrobotics/ign-math/pull-request/146)
+
+1. Added an assignment operator to the Frustum class.
+    * [Pull request 144](https://bitbucket.org/ignitionrobotics/ign-math/pull-request/144)
+
 1. Add static const variables as alternative to macros in Helpers.hh
     * [Pull request 137](https://bitbucket.org/ignitionrobotics/ign-math/pull-request/137)